pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-book-recommender"
version = "0.1.0"
description = "Portfolio-grade semantic book recommender with optional OpenAI/LangChain retrieval and Gradio UI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "python-dotenv>=1.0",
  "scikit-learn>=1.5",
]

[project.optional-dependencies]
app = [
  "gradio>=5.23",
]
llm = [
  "langchain>=0.3",
  "langchain-chroma>=0.2",
  "langchain-core>=0.3",
  "langchain-openai>=0.3",
  "torch>=2.5",
  "transformers>=4.48",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
